fix lambda coverage
